add a check for existing file and remove sudo

+++ init/30_install_davos.sh
@@ -1,3 +1,5 @@
 #!/bin/bash
 
-curl -o /app/davos.jar http://ci.linuxserver.io:8080/job/Davos/job/Release/ws/build/libs/davos-0.0.1-SNAPSHOT.jar
+[[ ! -f /app/davos.jar ]] && curl -o /app/davos.jar http://ci.linuxserver.io:8080/job/Davos/job/Release/ws/build/libs/davos-0.0.1-SNAPSHOT.jar
+
+chown abc:abc /app/davos.jar
